Форматирование
- Выделен компонент для карточки быстрой операции
- Выполнен мелкий рефакторинг
