+++ frontend/Money.Web/Pages/Operations/FastOperations.razor
@@ -25,182 +25,15 @@
 
 <MudDivider Class="mt-4 mb-4" />
 
-@if (_operations == null)
-{
-    return;
-}
-
 <MudGrid Spacing="2">
-    @foreach (FastOperation? operation in _operations)
-    {
-        // TODO: Подумать как более грамотно сделать
-        string className = operation.Category.OperationType.Id == 1 ? "expense-operation-card" : "income-operation-card";
-        if (operation.IsDeleted)
-        {
-            className = "deleted-operation-card";
-        }
-        <MudItem class="mb-1"
+    <Virtualize Context="operation"
+                Items="_operations">
+        <MudItem Class="mb-1"
                  lg="3"
                  md="4"
                  xs="12">
-            <MudCard Class="@className">
-                <MudCardHeader Style="padding: 4px 6px;">
-                    <MudTooltip RootClass="mud-card-header-text"
-                                Text="@operation.Name">
-                        <MudText Class="mud-card-header-text"
-                                 Typo="Typo.h6">
-                            @operation.Name
-                        </MudText>
-                    </MudTooltip>
-                </MudCardHeader>
-                <MudCardContent Style="padding: 4px 6px">
-                    <MudTooltip Arrow
-                                Placement="Placement.Right"
-                                Text="Категория">
-                        <MudText Typo="Typo.subtitle2">@operation.Category.Name</MudText>
-                    </MudTooltip>
-
-                    <MudDivider Class="mb-1" />
-
-                    <MudTooltip Arrow
-                                Placement="Placement.Right"
-                                Text="Сумма">
-                        <MudText Typo="Typo.body2">@operation.Sum</MudText>
-                    </MudTooltip>
-
-                    @if (string.IsNullOrEmpty(operation.Place) == false)
-                    {
-                        <MudDivider Class="mb-1" />
-
-                        <MudTooltip Arrow
-                                    Placement="Placement.Right"
-                                    Text="Место">
-                            <MudText Typo="Typo.body2">@operation.Place</MudText>
-                        </MudTooltip>
-                    }
-
-                    @if (string.IsNullOrEmpty(operation.Comment) == false)
-                    {
-                        <MudDivider Class="mb-1" />
-
-                        <MudTooltip Arrow
-                                    Placement="Placement.Right"
-                                    Text="Комментарий">
-                            <MudText Typo="Typo.body2">@operation.Comment</MudText>
-                        </MudTooltip>
-                    }
-                </MudCardContent>
-                <MudCardActions Style="padding: 4px 2px">
-                    @if (operation.Order != null)
-                    {
-                        <MudTooltip Text="Приоритет сортировки">
-                            <MudChip Class="sort-chip"
-                                     Icon="@Icons.Material.Filled.Sort"
-                                     Size="Size.Small"
-                                     Value="@operation.Order.ToString()" />
-                        </MudTooltip>
-                    }
-
-                    <MudSpacer />
-
-
-                    @if (!operation.IsDeleted)
-                    {
-                        <MudTooltip Text="Редактировать">
-                            <MudIconButton Color="Color.Warning"
-                                           Icon="@Icons.Material.Filled.Edit"
-                                           OnClick="() => Update(operation)"
-                                           Size="Size.Small" />
-                        </MudTooltip>
-                        <MudTooltip Text="Удалить">
-                            <MudIconButton Color="Color.Default"
-                            Icon="@Icons.Material.Filled.Delete"
-                            OnClick="() => Delete(operation)"
-                            Size="Size.Small" />
-                        </MudTooltip>
-                    }
-                    else
-                    {
-                        <MudTooltip Text="Восстановить">
-                            <MudIconButton Color="Color.Success"
-                                       Icon="@Icons.Material.Filled.Restore"
-                                       OnClick="() => Restore(operation)"
-                                       Size="Size.Small" />
-                        </MudTooltip>
-                    }
-                </MudCardActions>
-            </MudCard>
+            <FastOperationCard OnUpdate="Update"
+                               Operation="operation" />
         </MudItem>
-    }
+    </Virtualize>
 </MudGrid>
-
-
-<style>
-
-    .income-operation-card {
-        height: 100%;
-        transition: transform 0.2s, box-shadow 0.2s;
-        border: 1px solid rgba(76, 175, 80, 0.5);
-        border-radius: 8px;
-        background-color: rgba(76, 175, 80, 0.1);
-    }
-
-    .income-operation-card:hover {
-        transform: scale(1.03);
-        box-shadow: 0 4px 16px rgba(0, 0, 0, 0.2);
-    }
-
-    .income-operation-card .mud-card-header {
-        border-bottom: 1px solid rgba(76, 175, 80, 0.5);
-        background-color: rgba(76, 175, 80, 0.2);
-    }
-
-    .expense-operation-card {
-        height: 100%;
-        transition: transform 0.2s, box-shadow 0.2s;
-        border: 1px solid rgba(255, 193, 7, 0.5);
-        border-radius: 8px;
-        background-color: rgba(255, 193, 7, 0.1);
-    }
-
-    .expense-operation-card:hover {
-        transform: scale(1.03);
-        box-shadow: 0 4px 16px rgba(0, 0, 0, 0.2);
-    }
-
-    .expense-operation-card .mud-card-header {
-        border-bottom: 1px solid rgba(255, 193, 7, 0.5);
-        background-color: rgba(255, 193, 7, 0.2);
-    }
-
-    .mud-card-header-text {
-        max-width: 100%;
-        white-space: nowrap;
-        overflow: hidden;
-        text-overflow: ellipsis;
-    }
-
-    .deleted-operation-card {
-        height: 100%;
-        transition: transform 0.2s, box-shadow 0.2s;
-        border: 1px solid rgba(128, 128, 128, 0.5);
-        border-radius: 8px;
-        background-color: rgba(128, 128, 128, 0.1);
-    }
-
-    .deleted-operation-card:hover {
-        transform: scale(1.02);
-        box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-    }
-
-    .deleted-operation-card .mud-card-header {
-        border-bottom: 1px solid rgba(128, 128, 128, 0.5);
-        background-color: rgba(128, 128, 128, 0.2);
-    }
-
-    .sort-chip {
-        color: white;
-        background-color: #1976d2;
-    }
-
-</style>
