+++ frontend/Money.Web/Pages/Operations/FastOperations.razor.cs
@@ -1,33 +1,20 @@
 ﻿using Microsoft.AspNetCore.Components;
-using Money.ApiClient;
 using Money.Web.Components.FastOperations;
 
 namespace Money.Web.Pages.Operations;
 
 public partial class FastOperations
 {
-    private List<Category>? _categories;
-
-    private List<FastOperation>? _operations;
-
-    [Inject]
-    private MoneyClient MoneyClient { get; set; } = null!;
-
-    [Inject]
-    private CategoryService CategoryService { get; set; } = null!;
+    private List<FastOperation> _operations = [];
 
     [Inject]
     private FastOperationService FastOperationService { get; set; } = null!;
-
-    [Inject]
-    private ISnackbar SnackbarService { get; set; } = null!;
 
     [Inject]
     private IDialogService DialogService { get; set; } = null!;
 
     protected override async Task OnInitializedAsync()
     {
-        _categories = await CategoryService.GetAllAsync();
         _operations = await FastOperationService.GetAllAsync();
     }
 
@@ -52,33 +39,6 @@ public partial class FastOperations
     private Task<FastOperation?> Update(FastOperation fastOperation)
     {
         return ShowDialog("Обновить", fastOperation);
-    }
-
-    private Task Delete(FastOperation fastOperation)
-    {
-        return Modify(fastOperation, MoneyClient.FastOperation.Delete, true);
-    }
-
-    private Task Restore(FastOperation fastOperation)
-    {
-        return Modify(fastOperation, MoneyClient.FastOperation.Restore, false);
-    }
-
-    private async Task Modify(FastOperation fastOperation, Func<int, Task<ApiClientResponse>> action, bool isDeleted)
-    {
-        if (fastOperation.Id == null)
-        {
-            return;
-        }
-
-        ApiClientResponse result = await action(fastOperation.Id.Value);
-
-        if (result.GetError().ShowMessage(SnackbarService).HasError())
-        {
-            return;
-        }
-
-        fastOperation.IsDeleted = isDeleted;
     }
 
     private async Task<FastOperation?> ShowDialog(string title, FastOperation fastOperation)
